Split runtime world matrix logic
